Removed unnecessary casting

+++ Test/gsapTests/ModelBasedPrognoserTests.cpp
@@ -85,8 +85,8 @@ void testWithMockModel() {
     result = mbp.step(data);
     
     Assert::AreEqual(result.getEvents().size(), 1);
-    Assert::AreEqual((double) result.getEvents()[0].getState()[0].get(), 1, 1e-6);
-    Assert::AreEqual((double) result.getEvents()[0].getStartTime().get(), 1.5, 1e-6);
+    Assert::AreEqual(result.getEvents()[0].getState()[0].get(), 1, 1e-6);
+    Assert::AreEqual(result.getEvents()[0].getStartTime().get(), 1.5, 1e-6);
     Assert::AreEqual(result.getSystemTrajectories().size(), 0);
     
     // No time passed
